github & helpers: split Lando-specific GitHub logic (bug 2057141)
This commit separates Lando- and Django-specific GitHub adapters from
the project-agnostic GitHub logic. This mainly moves methods to new
files, though some tidbits of logic are move to the call-sites, with the
goal to not import any project module from the util.github component.

This does the same for tests, separating them to match the split, and
moving GitHub fixtures to a dedicated conftest, which is explicitly
re-imported.
