+++ conftest.py
@@ -1,0 +1,4 @@
+# pytest_plugins needs to be defined in the root conftest.py.
+pytest_plugins = [
+    "lando.utils.tests.conftest_github",
+]
