+++ src/lando/utils/github.py
@@ -1,33 +1,36 @@
+"""This file contains project-independent GitHub utils.
+
+WARNING: Do not import and lando- or django-namespaced modules here.
+
+"""
+
 import asyncio
-import functools
 import hashlib
 import hmac
-import io
-import json
 import logging
 import math
 import re
 from collections import Counter
 from collections.abc import Callable, Iterator
+from dataclasses import dataclass
 from datetime import datetime
 from enum import Enum
 from itertools import count
-from json.decoder import JSONDecodeError
 
 import requests
-from django.conf import settings
-from django.core.handlers.wsgi import WSGIRequest
-from django.http import HttpResponse
-from django.views import View
 from simple_github import AppAuth, AppInstallationAuth
-from typing_extensions import override
 
-from lando.main.models.configuration import ConfigurationKey, ConfigurationVariable
-from lando.main.scm.helpers import PatchHelper, PatchHelperMetadata
-from lando.utils.cache import cache_method
-from lando.utils.const import URL_USERINFO_RE
+from .cache import cache_method
+from .const import URL_USERINFO_RE
 
 logger = logging.getLogger(__name__)
+
+
+@dataclass
+class settings:
+    GITHUB_APP_ID: str | None = None
+    GITHUB_APP_PRIVKEY: str | None = None
+    HTTP_USER_AGENT: str = "Mozilla-GitHub-Util/v0.0.0-pre"
 
 
 PR_DELIMITER = (
@@ -632,18 +635,8 @@ class PullRequest:
 
     @property
     def reviews_summary(self) -> dict[str, str]:
-        """Get a simple dict of reviewers and the state of their review.
-
-        The reviewers GitHub usernames are mapped to nicks, if available in the
-        GITHUB_REVIEWERS_MAP ConfigurationVariable.
-        """
-        summary = {review["user"]["login"]: review["state"] for review in self.reviews}
-
-        reviewer_map = ConfigurationVariable.get(
-            ConfigurationKey.GITHUB_REVIEWERS_MAP, {}
-        )
-
-        return {reviewer_map.get(u, u): summary[u] for u in summary}
+        """Get a simple dict of reviewers and the state of their review."""
+        return {review["user"]["login"]: review["state"] for review in self.reviews}
 
     @property
     @pr_cache_method
@@ -709,79 +702,6 @@ class PullRequest:
         }
 
 
-class PullRequestPatchHelper(PatchHelper):
-    """A PatchHelper-like wrapper for GitHub pull requests.
-
-    Due to the nature of pull requests, it only implement the data-getting
-    functionality, and doesn't implement the input and output methods.
-    """
-
-    _diff: str
-
-    _author_name: str
-    _author_email: str
-    _pr: PullRequest
-
-    def __init__(self, pr: PullRequest):
-        super().__init__()
-
-        self._pr = pr
-
-        self._diff = pr.diff
-
-        author_name, author_email = self._pr.author
-
-        self.headers = {
-            "date": self._get_timestamp_from_github_timestamp(pr.updated_at),
-            "from": f"{author_name} <{author_email}>",
-            "subject": pr.title,
-        }
-
-        self.metadata = PatchHelperMetadata()
-
-    @classmethod
-    def _get_timestamp_from_github_timestamp(cls, timestamp: str) -> str:
-        timestamp_datetime = datetime.fromisoformat(timestamp)
-        return str(math.floor(timestamp_datetime.timestamp()))
-
-    @classmethod
-    def from_string_io(cls, string_io: io.StringIO) -> "PatchHelper":
-        """Implement the PatchHelper interface; not relevant for GitHub PRs."""
-        raise NotImplementedError("`from_string_io` not implemented.")
-
-    @classmethod
-    def from_bytes_io(cls, bytes_io: io.BytesIO) -> "PatchHelper":
-        """Implement the PatchHelper interface; not relevant for GitHub PRs."""
-        raise NotImplementedError("`from_bytes_io` not implemented.")
-
-    def get_commit_description(self) -> str:
-        """Returns the commit description."""
-        return self.get_header("subject")
-
-    @override
-    def get_diff(self) -> str:
-        """Return the patch diff.
-
-        WARNING: As of 2025-10-13, this doesn't include any binary data.
-        """
-        return self._diff
-
-    @override
-    def write(self, f: io.StringIO):
-        """Implement the PatchHelper interface; not relevant for GitHub PRs."""
-        raise NotImplementedError("`from_bytes_io` not implemented.")
-
-    @override
-    def parse_author_information(self) -> tuple[str, str]:
-        """Return the author name and email from the patch."""
-        return self._pr.author
-
-    @override
-    def get_timestamp(self) -> str:
-        """Return an `hg export` formatted timestamp."""
-        return self.get_header("date")
-
-
 def verify_github_signature(hmac_secret: str, payload: bytes, signature: str) -> bool:
     """Verify the provided signature of the payload using the given hmac_secret."""
     # NOTE: this was adapted from code provided in GitHub as an example.
@@ -793,22 +713,3 @@ def verify_github_signature(hmac_secret: str, payload: bytes, signature: str) ->
     )
     expected_signature = "sha256=" + hash_object.hexdigest()
     return hmac.compare_digest(expected_signature, signature)
-
-
-def ignore_bot_sender(post: Callable) -> Callable:
-    """Decorator that drops requests that originate from bots."""
-
-    @functools.wraps(post)
-    def _post(view: View, request: WSGIRequest, *args, **kwargs) -> HttpResponse:
-        """Drop the request if a bot triggered the original webhook."""
-        BOT_SENDER_TYPE = "Bot"
-        try:
-            sender_type = json.loads(request.body)["sender"]["type"]
-        except (JSONDecodeError, KeyError, ValueError, TypeError):
-            pass
-        else:
-            if sender_type == BOT_SENDER_TYPE:
-                return HttpResponse(status=202)
-        return post(view, request, *args, **kwargs)
-
-    return _post
